Fixing bug 16
This one was really stupifying. I ended up copping out and using
innerHTML because the manifested problem was so bizarre. If there
were multiple graphs loaded from an [All] or playlist and you deleted
any graph and then hit the 'delete all' trashcan, all but one graph
would be deleted and the app would be sorta hozed. I debugged it for
hours and ended up giving up because the loop would simply not go as
far as it should. Bummer.