here's a synopsis: prototype SUCKS
So Prototype's Enumerable ".each" uses a count of the elements of an array rather than the actual keys of the array!?!? What's the point of that? I *think* the only problematic arrays have been acted on by 'delete' and G.graphs is the only array I use 'delete' on, so I replaced all of the arrays that loop over G.graphs with the entirely crappy loops that have to check the type of each key!

Remind me to rewrite this whole thing in jQuery.(Assuming it doesn't have the same problem.)